Update to spanish the db and sql consulst, divided functions in inventory, medic and users
