+++ includes/inventory/itemsTable.php
@@ -1,6 +1,6 @@
 <?php
 $conn = $GLOBALS['conn'];
-$sql = "SELECT * FROM inventory";
+$sql = "SELECT * FROM inventario";
 $result = mysqli_query($conn, $sql);
 $row_inventory = mysqli_fetch_all($result, MYSQLI_ASSOC);
 
@@ -9,10 +9,10 @@ $inventoryItems = array();
 foreach ($row_inventory as $row) {
     // Crear un objeto para cada elemento del inventario
     $item = array(
-        'nameItem' => $row["nameItem"],
-        'descriptionItem' => $row["descriptionItem"],
-        'countItem' => $row["countItem"],
-        'idItem' => $row["idItem"]
+        'nameItem' => $row["nombreArticulo"],
+        'descriptionItem' => $row["descripcionArticulo"],
+        'countItem' => $row["cantidadArticulo"],
+        'idItem' => $row["idArticulo"]
     );
 
     $inventoryItems[] = $item;
