+++ includes/medic/patientsTable.php
@@ -1,6 +1,6 @@
 <?php
 $conn = $GLOBALS['conn'];
-$sql = "SELECT * FROM patients";
+$sql = "SELECT * FROM pacientes";
 $result = mysqli_query($conn, $sql);
 $row_inventory = mysqli_fetch_all($result, MYSQLI_ASSOC);
 
@@ -8,12 +8,12 @@ $inventoryItems = array();
 
 foreach ($row_inventory as $row) {
     $patient = array(
-        'PatienName' => $row["patientName"],
-        'patientLastName' => $row["patientLastName"],
-        'patientCd' => $row["patientCd"],
-        'patientDOB' => $row["patientDOB"],
-        'patientGender' => $row["patientGender"],
-        'idPatient' => $row["idPatient"]
+        'PatienName' => $row["nombrePaciente "],
+        'patientLastName' => $row["apellidoPaciente"],
+        'patientCd' => $row["cedulaPaciente"],
+        'patientDOB' => $row[" 	fechaNacimientoPaciente"],
+        'patientGender' => $row["generoPaciente"],
+        'idPatient' => $row["idPaciente"]
 
     );
